+++ NEWS.md
@@ -7,6 +7,8 @@
   
 * `across()` handles data frames with 0 columns (#5523). 
 
+* `mutate()` always keeps grouping variables, unconditional to `.keep=` (#5582).
+
 # dplyr 1.0.2
 
 * Fixed `across()` issue where data frame columns would mask objects referred to